+++ vm_provision/provision-sailsjs.sh
@@ -1,4 +1,7 @@
 #install sailsjs
-echo "----- Provision: Installing sailsjs globally"
-sudo npm -g install sails
+if [ ! -f /usr/bin/sailsjs ]; then
+	echo "----- Provision: Installing sailsjs globally"
+	sudo npm -g install sails
+fi
+
